requirements.txt
flask==3.0.0
requests==2.31.0
lxml==5.1.0
google-generativeai==0.3.2
fpdf==1.7.2